Split into lib/bin
For early development, everything was in `main.rs`. This commit moves to
something more idiomatic; we have all of our logic in a `lib.rs`, and
then do argument parsing and validation in `main`, and delegate to `lib`
for all the logic.

Fixes #14.
